moving types to type.ts
