+++ nexus-sdk-js/src/View/SparqlView/index.ts
@@ -1,39 +1,7 @@
 import { httpPost } from '../../utils/http';
 import { SparqlViewQueryException } from './exceptions';
 import { getSparqlView } from '../utils';
-
-interface SparqlQuery {
-  content: string;
-}
-
-export interface SparqlViewResponse {
-  '@id': string;
-  '@type': string[];
-  _uuid: string;
-  _rev: number;
-  _deprecated: boolean;
-}
-
-// based on: https://www.w3.org/TR/sparql11-results-json/
-// describes a SPARQL Query Results JSON
-export interface SparqlViewQueryResponse {
-  head: {
-    vars: string[];
-    link?: string | string[];
-  };
-  results?: {
-    bindings: {
-      [key: string]: {
-        type: string;
-        value: any;
-        datatype?: string;
-        'xml:lang'?: string;
-        [attribute: string]: any;
-      };
-    }[];
-  };
-  boolean?: boolean;
-}
+import { SparqlViewResponse, SparqlViewQueryResponse } from './types';
 
 export default class SparqlView {
   id: string;
